chore(bench): extract shared fixtures to reduce duplication
Create __test__/bench-fixtures.ts with shared test data (string pairs,
candidate arrays, dataset generators) used across benchmark files.
Removes duplication of pairs array from 3 files and dataset generators
from search.bench.ts.

Add bench-fixtures.ts to CodSpeed workflow path triggers.

Closes #373
