+++ .github/workflows/codspeed.yml
@@ -6,6 +6,7 @@ on:
     paths:
       - "crates/**"
       - "__test__/**/*.bench.ts"
+      - "__test__/bench-fixtures.ts"
       - "vitest.config.mts"
       - ".github/workflows/codspeed.yml"
   pull_request:
@@ -13,6 +14,7 @@ on:
     paths:
       - "crates/**"
       - "__test__/**/*.bench.ts"
+      - "__test__/bench-fixtures.ts"
       - "vitest.config.mts"
       - ".github/workflows/codspeed.yml"
   workflow_dispatch:
