+++ __test__/batch.bench.ts
@@ -5,24 +5,7 @@ import {
   normalizedLevenshtein,
   normalizedLevenshteinMany,
 } from '../index.js';
-
-// --- Many candidates (1-to-N comparison) ---
-
-const manyCandidates = Array.from({ length: 1_000 }, (_, i) => {
-  const words = [
-    'kitten',
-    'sitting',
-    'saturday',
-    'sunday',
-    'hello',
-    'world',
-    'fuzzy',
-    'search',
-    'match',
-    'distance',
-  ];
-  return `${words[i % words.length]}${i}`;
-});
+import { manyCandidates } from './bench-fixtures.js';
 
 describe('Levenshtein Distance — Many (1K candidates)', () => {
   bench('rapid-fuzzy (many)', () => {
