+++ __test__/distance.bench.ts
@@ -1,15 +1,6 @@
 import { bench, describe } from 'vitest';
 import { damerauLevenshtein, hamming, levenshtein, levenshteinBatch } from '../index.js';
-
-// Test data — realistic string pairs of varying length and similarity
-const pairs: [string, string][] = [
-  ['kitten', 'sitting'],
-  ['saturday', 'sunday'],
-  ['rosettacode', 'raisethysword'],
-  ['pneumonoultramicroscopicsilicovolcanoconiosis', 'ultramicroscopically'],
-  ['the quick brown fox jumps over the lazy dog', 'the fast brown fox leaps over the lazy dog'],
-  ['abcdefghijklmnopqrstuvwxyz', 'zyxwvutsrqponmlkjihgfedcba'],
-];
+import { equalLengthPairs, pairs } from './bench-fixtures.js';
 
 describe('Levenshtein Distance', () => {
   bench('rapid-fuzzy', () => {
@@ -30,16 +21,6 @@ describe('Damerau-Levenshtein', () => {
     }
   });
 });
-
-// Equal-length pairs for Hamming distance
-const equalLengthPairs: [string, string][] = [
-  ['karolin', 'kathrin'],
-  ['saturday', 'sunturdy'],
-  ['abcdefgh', 'abcdefgz'],
-  ['10101010', '01010101'],
-  ['the quick brown fox jumps', 'the swift brown fox leaps'],
-  ['abcdefghijklmnopqrstuvwxyz', 'zyxwvutsrqponmlkjihgfedcba'],
-];
 
 describe('Hamming Distance', () => {
   bench('rapid-fuzzy', () => {
