+++ __test__/search.bench.ts
@@ -1,135 +1,6 @@
 import { bench, describe } from 'vitest';
-
 import { closest, FuzzyIndex, search } from '../index.js';
-
-// --- Test data ---
-
-// Small dataset (typical autocomplete)
-const smallItems = [
-  'apple',
-  'banana',
-  'grape',
-  'orange',
-  'pineapple',
-  'mango',
-  'strawberry',
-  'blueberry',
-  'raspberry',
-  'watermelon',
-  'kiwi',
-  'peach',
-  'plum',
-  'cherry',
-  'lemon',
-  'lime',
-  'coconut',
-  'avocado',
-  'papaya',
-  'guava',
-];
-
-// Medium dataset (file finder, ~1K items)
-const mediumItems = Array.from({ length: 1_000 }, (_, i) => {
-  const dirs = ['src', 'lib', 'test', 'docs', 'utils', 'components', 'hooks', 'services'];
-  const exts = ['.ts', '.js', '.tsx', '.jsx', '.json', '.md'];
-  const names = [
-    'index',
-    'main',
-    'utils',
-    'helper',
-    'config',
-    'types',
-    'schema',
-    'handler',
-    'service',
-    'controller',
-  ];
-  const dir = dirs[i % dirs.length];
-  const name = names[i % names.length];
-  const ext = exts[i % exts.length];
-  return `${dir}/${name}${Math.floor(i / 10)}${ext}`;
-});
-
-// Large dataset (~10K items)
-const largeItems = Array.from({ length: 10_000 }, (_, i) => {
-  const words = [
-    'async',
-    'await',
-    'function',
-    'class',
-    'interface',
-    'type',
-    'export',
-    'import',
-    'const',
-    'let',
-    'return',
-    'promise',
-    'observable',
-    'subscriber',
-    'handler',
-    'middleware',
-    'controller',
-    'service',
-    'repository',
-    'factory',
-  ];
-  return `${words[i % words.length]}_${words[(i * 7) % words.length]}_${i}`;
-});
-
-// Extra-large dataset (~50K items)
-const xlargeItems = Array.from({ length: 50_000 }, (_, i) => {
-  const words = [
-    'async',
-    'await',
-    'function',
-    'class',
-    'interface',
-    'type',
-    'export',
-    'import',
-    'const',
-    'let',
-    'return',
-    'promise',
-    'observable',
-    'subscriber',
-    'handler',
-    'middleware',
-    'controller',
-    'service',
-    'repository',
-    'factory',
-  ];
-  return `${words[i % words.length]}_${words[(i * 7) % words.length]}_${i}`;
-});
-
-// 100K dataset (scaling benchmark)
-const hugeItems = Array.from({ length: 100_000 }, (_, i) => {
-  const words = [
-    'async',
-    'await',
-    'function',
-    'class',
-    'interface',
-    'type',
-    'export',
-    'import',
-    'const',
-    'let',
-    'return',
-    'promise',
-    'observable',
-    'subscriber',
-    'handler',
-    'middleware',
-    'controller',
-    'service',
-    'repository',
-    'factory',
-  ];
-  return `${words[i % words.length]}_${words[(i * 7) % words.length]}_${i}`;
-});
+import { hugeItems, largeItems, mediumItems, smallItems, xlargeItems } from './bench-fixtures.js';
 
 // --- Pre-initialize search instances ---
