+++ __test__/similarity.bench.ts
@@ -1,15 +1,6 @@
 import { bench, describe } from 'vitest';
 import { jaro, jaroWinkler, normalizedLevenshtein, sorensenDice } from '../index.js';
-
-// Test data — realistic string pairs of varying length and similarity
-const pairs: [string, string][] = [
-  ['kitten', 'sitting'],
-  ['saturday', 'sunday'],
-  ['rosettacode', 'raisethysword'],
-  ['pneumonoultramicroscopicsilicovolcanoconiosis', 'ultramicroscopically'],
-  ['the quick brown fox jumps over the lazy dog', 'the fast brown fox leaps over the lazy dog'],
-  ['abcdefghijklmnopqrstuvwxyz', 'zyxwvutsrqponmlkjihgfedcba'],
-];
+import { pairs } from './bench-fixtures.js';
 
 describe('Normalized Similarity', () => {
   bench('rapid-fuzzy (normalizedLevenshtein)', () => {
